+++ tests/scalers/rabbitmq/rabbitmq_queue_amqp/rabbitmq_queue_amqp_test.go
@@ -32,8 +32,8 @@ var (
 	queueName        = "hello"
 	user             = fmt.Sprintf("%s-user", testName)
 	password         = fmt.Sprintf("%s-password", testName)
-	vhost            = fmt.Sprintf("%s-vhost", testName)
-	connectionString = fmt.Sprintf("amqp://%s:%s@rabbitmq.%s.svc.cluster.local/%s", user, password, rmqNamespace, vhost)
+	vhost            = "/"
+	connectionString = fmt.Sprintf("amqp://%s:%s@rabbitmq.%s.svc.cluster.local", user, password, rmqNamespace)
 	messageCount     = 100
 )